Added a CVS snapshot of SDL_ttf in the wesnoth tree Made wesnoth link statically to those SDL_ttf files Added checking for libfreetype in the autotools scripts.
Warning: changes made to the autotools script! they work for me, but need testing and reviewing!

The main reason for that is that SDL_ttf 2.0.6 still has fatal
memory-corruption bugs, fixed in CVS. This fix looks like to prevent manu
segfaults for some configurations using the French translation.

This fix may be reverted when SDL_ttf 2.0.7 is released (or not).


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@4463 75b3775d-3801-0410-9e12-e2e43121521d