fixes bug #5356 (player gets kicked when all slots are filled up). The behaviour is now this: The player that first presses "join game" (more precisely whose join-message gets processed first on the server) is in. Everyone else will only be able to join as observer if no more free sides are left. This happens immediately after joining, no matter how much time is spent in the faction choosing dialog.
I also reworked some of the code by introducing a new observer list for a game (so far, player_ had both players and observers in it). It helps a lot with the bug and i feel it is much less confusing.
I also hope that problems with corrupted observer lists ingame are fixed now and that you should not get an empty lobby if you get kicked out of the game (by whatever reason).

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@11263 75b3775d-3801-0410-9e12-e2e43121521d