Add a new field to the game state structure: completion, which may have the values 'running', 'victory', or 'defeat'.  It is set by play_game() and check_victory().  It is loaded and saved.
Presently this is not yet used (though I have tested that it is set properly).
So there are no behavioral changes in the game yet.

What it does is give us a way of setting victory/defeat status that isn't 
tied to throwing an end-of-scenario exception.  This is a first step to 
allowing observer-like game browsing after victory/defeat notification,
with the user choosing to throw end-of-scenario after that.

Doing this required passing a game-state reference into two classes 
that didn't have it before, ai_interface and mouse_handler.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@18277 75b3775d-3801-0410-9e12-e2e43121521d