MSVC 6 and 9 cleanup.
Now that MSVC 6 is no longer supported start to remove some helper code to work
around problems in that compiler. The first step is to comment out the code and
replace it with new helper code that uses the standard. The second step is to
remove the helper code and fix all callers.

Patch 1/3 for step 2: Clear global.hpp, but add some MSVC9 workarounds.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@28937 75b3775d-3801-0410-9e12-e2e43121521d