Add a new title screen.
Now that all 'stable' --new-widgets code has been moved to the default way
starting with --new-widgets is no longer needed unless you want to develop for
it. So the new proof-of-concept title screen can be committed. Only has the
basic buttons which work, but needs some cleanup up, both in looks and code
working.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@28945 75b3775d-3801-0410-9e12-e2e43121521d