Change the text storage type to std::string.
I thought the t_string had some magic to automatically change it's text
when the translation changes. That's not the case so no longer need it
to be a t_string. Also future changes to the text widgets will get
problems since I want to return references which might be temporaries
due to the conversions in t_string so rather avoid that as well.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@29009 75b3775d-3801-0410-9e12-e2e43121521d