Add a new debug class for the widgets.
This debug option can output dot files which can be transformed to images so
it's easier to get debug info about the widgets in a window. This function is
enabled by the --enable-debug-window-layout configure switch. It has not been
implemented in the other build systems yet. (Not sure whether or not it's
needed.)


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@30699 75b3775d-3801-0410-9e12-e2e43121521d