Improved the map scrolling feature.
After r30918 fixed the scrolling outside the window, several people noticed the
threshold of 5 pixels is rather small. Added a hidden preference named
'scroll_threshold' which is used to determine the number of pixels. The
threshold also defaults to 10 instead of 5, more testing is needed to get a
good default value.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@31339 75b3775d-3801-0410-9e12-e2e43121521d