Stricter check before sending events to widgets.
When a scrollbar button was clicked and that click disabled the button
since it reached the end of the range, the next mouse move event would
send a mouse leave event to that widget, which reenabled the widget.
Also added the check to the entering code (not sure whether it can be
triggered, but rather save than sorry).

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@31914 75b3775d-3801-0410-9e12-e2e43121521d