Fix a crash with a long text.
At the moment the scroll_label is the label to be used if wrapping is required.
The dialog looks a bit ugly due to the fact that the scroll_label can't
optimize it's size yet. Also the alignment looks a bit odd, but I'll leave it
to Baufo to tune that part.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@32429 75b3775d-3801-0410-9e12-e2e43121521d