Removed the gamma correction option.
The code caused quite some crashes and we're not convinced the option is really
that useful. The code has only been disabled to evaluate how many people will
miss the feature. If really not useful the code needs to be removed and the
preferences dialog display tab needs a cleanup to look good without this
option (the other options need to be moved upwards).


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@32813 75b3775d-3801-0410-9e12-e2e43121521d