Don't use a scrollbar in an invisible widget.
AI0867 discovered the following problem:
start the test scenario, move a unit to "open sesame", move a unit to
"math test", enter a number longer than the input box, select "ints &
floats".

The game ended with an assertion failure due to using a scrollbar on an
invisible widget. Added a minimal fix, this needs to be fixed cleaner
post 1.6.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@33437 75b3775d-3801-0410-9e12-e2e43121521d