Pango-enable message= attributes within [message].
This patch includes (a) switching on this feature in the C core (and
diabling grotty old Wesnoth markup in this context), (b) an upgrade to
wmllint to do most of these conversions automatically (e.g. in TROW),
(c) hand-fixes for some unual cases (in the Tutorial), and (d) a test
of Pango <i> (in THOT).


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@34263 75b3775d-3801-0410-9e12-e2e43121521d