Add a helper header for a small test.
Quoting the header:
This file is being used for a small experiment in which some private
functions of tgrid are no longer in tgrid but moved in a friend class with
static functions. The goal is to have less header recompilations, when
there's a need to add or remove a private function.
Also non-trivial functions like 'const foo& bar() const' and 'foo& bar()'
are wrapped in a template to avoid code duplication (for typing not for the
binary) to make maintenance easier.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@35038 75b3775d-3801-0410-9e12-e2e43121521d