Add a new generic logging header for gui2.
r35170 removed the specialized log domains and the macros to make logging
easier. I added them so I don't need to add the same items to the gui files all
the time. The auxiliary directory was alread longer planned so some file in the
widgets directory which are no widgets can be moved there, this will be done
later.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@35176 75b3775d-3801-0410-9e12-e2e43121521d