Added a minimal implementation for the tooltips.
Since the tooltips are currently only used for the multiplayer menu the
have been impemented so they work there.
Post 1.8 it will be implemented better but it might be that the tooltips
then become separate floating windows. Since that's not sure added the
minimal version for 1.8 and look further after that version has been
shipped.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@39327 75b3775d-3801-0410-9e12-e2e43121521d