Add better protection for images not fitting.
Some UMC uses listbox rowes with less than 20 pixes which caused images
to be resized to x - 20 which is negative and resulted into huge
positive numbers.

Modified the WML to return 0 in these cases and also added protection in
the engine to ignore those images. (The latter is needed since people
can create their own themes which might have this problem as well.)

Fixes bug #14525 and probably bug #14002.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@39557 75b3775d-3801-0410-9e12-e2e43121521d