Add a new repeating button widget.
A repeating button is button when pressed down will execute an action, if the
user keeps it pressed down for a longer time it will repeat this action until
released.

The main place where the widget was missed is the MP lobby, I haven't tested it
there much since the lobby is still to slow. This means once the lobby is fast
enough it needs more tuning.

The widget hasn't been wired in the scrollbars, which will be done in the next
commit.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@40229 75b3775d-3801-0410-9e12-e2e43121521d