Fixes some listbox assertion failures.
Ilor discovered them when showing and hiding items in the MP dialog.
This change also moves the sizing burden to the listbox, this was
documented in the contract but not done yet.

This allows the multiple item version of set_row_shown() to update the
size once instead of for every call to generator function.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@40531 75b3775d-3801-0410-9e12-e2e43121521d