Optimize the lobby speed.
Try to resize the content of the listboxes with their current content,
only if that fails invalidate the entire window layout.

Since the playerlist consists of three fixed size listboxes the
invalidation still occurs rather often. Especially when a game is
selected the resize happens often since most of the time the number of
people slotted in different listboxes are different, causing a resize.

Unfortunately that's hard to fix in the current setup, the fix would be
to add a proper treeview widget. Then all players are in one widget and
moving an item in one widget should be easier to handle without a window
invalidation.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@40723 75b3775d-3801-0410-9e12-e2e43121521d