Proof of concept to translate network errors.
This way network errors can be shown localized, after the string freeze
is lifted it should be possible to translate more strings. Strings with
variables can't be translated this way.

Changed the string in network.cpp to one which is already in the
translation database (so this commit shouldn't break the string freeze).
Tested with the German translation and the connection error is now shown
in German.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@41072 75b3775d-3801-0410-9e12-e2e43121521d