Fix some linking issues.
While testing clang ran into a linker issue in cmake. This patch reverts
r48086 and fixes the issue properly for both cmake and scons.

The patch has been tested with r48086 using autotools, cmake and scons
and all systems worked propely. So if somebody cares it would be
possible to re-add autotools.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@49313 75b3775d-3801-0410-9e12-e2e43121521d