Re-rationization of terrain names, part 1
Eleazar's "rationalization" of terrain names in r47151, r47154 left quite a bunch of terrains with editor_name attributes but no name attribute. While it may theoretically make sense to inherit the name attribute from an aliased terrain, this creates new problems of its own when a terrain is aliased to more than one type (which one to choose for name inheritance?). In my opinion it is best to solve this in WML and not add more layers (heh) of convolutedness to the terrain engine.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@50047 75b3775d-3801-0410-9e12-e2e43121521d