Reverted r51250 and r51240.
thonsew: The lua interface is not your playground. This is not the way I want the lua files or its C++ interface to be modified. You are invited to talk
with me on IRC.

some of the main reasons:
-Your revisions contain lots of unneccessary changes, some of which break existing lua or wml without a reason. Do you ever look at the diffs ?
-Your revisions pack a lot of unrelated stuff together into one.

r51250 makes sense in some parts and is what you should have done in the first place.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@51255 75b3775d-3801-0410-9e12-e2e43121521d