Apply "Setting to ignore encountered_units" (patch #2940 by negusnyul)
The preferences entry name and tooltip are not the same as the patch in
the tracker, as we discussed on IRC that this doesn't exclusively affect
mainline units nor does it need to. I also changed the tooltip wording
to fit as much as possible in my 1280x800 screen without changing the
font size (works on 800x480 with a smaller font size).

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@51516 75b3775d-3801-0410-9e12-e2e43121521d