accept only non-empty variable names in [set_variable(s)] and wesnoth.set_variable
This supposedly common problem (value of a set variable composed
of the values of other variables which perhaps don't exist or are empty)
causes corrupted savegames and makes :inspect fail. Hard to track on
wml side; there are probably a more ways of corrupting saves via empty
key or tag names.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@52481 75b3775d-3801-0410-9e12-e2e43121521d