Restore leader unit image in Load Game dialog (bug #18683)
(Candidate for 1.10 branch.)

This functionality is back with a slight change to r47758 in order to
left-align the map snapshot only when the leader image cannot be loaded.
We also no longer spam stderr with errors about missing leader images
(which can be caused by unloaded campaigns or eras).

Anonymissimus pointed out this functionality was still found in the GUI2
version of the dialog (--new-widgets). Whoever last changed the saved
games summary code probably only bothered to update this experimental
code without fixing the normal GUI1 dialog. As I pointed out in the
tracker, the layout code was still in place.

As a compromise for the time being, we always assume the "magenta"
palette for TC recoloring now. This doesn't seem to be a problem since
the old code already made an annoying assumption of side 1 using the 1
(red) color range instead of permiting the use of the side.color value,
or reflecting the currently playing side number at the time of the save
generation.

I may later extend the saved games summary to include additional
information to get rid of the magenta and side 1 limitations.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@52921 75b3775d-3801-0410-9e12-e2e43121521d