addon/mg: Better fix for builds with non-GNU C++ STL than r53210
The point of the at() member call was to throw an exception if the
access was invalid. By using find() without checking, we could end up
dereferencing the end iterator of the addons_list. Thus, a helper
function has been added now to use an assertion to protect against that
hypothetical situation early enough.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@53213 75b3775d-3801-0410-9e12-e2e43121521d