addon/mg: Make get_addon_version_info() return by value, not reference
The return statement makes use of the ternary ?: operator to select the
referenced object, and this construct in assignments has been found to
cause undefined behavior scenarios before (r34865, r34871, r34872), so
it's probably better to err on the side of caution here. The
version_info class instance copy overhead should be near-negligible.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@53411 75b3775d-3801-0410-9e12-e2e43121521d