gui2/teditor_settings: Use a cheaper, less CPU-intensive solution to refresh tiles after changing ToD settings
Instead of refreshing the entire display, this solution simply
invalidates all gamemap tiles and requests them to be redrawn, saving a
lot of CPU time and making continuous operations (i.e. dragging one of
the sliders around) smoother. We still need to redraw the dialog,
though, and there's a hack involved to avoid a floating slice of it
appearing in the middle of the process.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@53593 75b3775d-3801-0410-9e12-e2e43121521d